+++ Assets/Scripts/Boss/GenericAttack.cs
@@ -33,7 +33,7 @@ public class GenericAttack : MonoBehaviour
 
     private void OnTriggerEnter2D(Collider2D collision) {
         if (collision.CompareTag("Player")) {
-            collision.GetComponent<PlayerMovement>().DamageOverTimeRoutine(value, duration);
+            collision.GetComponent<PlayerHealth>().DamageOverTimeRoutine(value, duration);
             Destroy(gameObject);
         }
     }
