+++ Assets/Scripts/Player/PlayerMovement.cs
@@ -5,15 +5,8 @@ using System.Collections;
 public class PlayerMovement : MonoBehaviour
 {
     private Rigidbody2D body;
-    private Coroutine healRoutine;
-    private float healRoutineTimeLeft;
-    private Coroutine damageRoutine;
-    private float damageRoutineTimeLeft;
     public float speed = 5f;
     public float jumpForce = 5f;
-    public int startingHealth = 100;
-    public int health = 100;
-    public int lives = 3;
 
     private Animator animator;
     private SpriteRenderer spriteRenderer;
@@ -53,62 +46,6 @@ public class PlayerMovement : MonoBehaviour
         if (body.linearVelocityX > 0)
         {
             spriteRenderer.flipX = false;
-        }
-    }
-
-    public void ModifyHealth(int value) {
-        health += value;
-
-        if (health <= 0) {
-            lives--;
-            health = startingHealth;
-            transform.position = new Vector2(0,0);
-        } else if (health >= startingHealth) {
-            health = startingHealth;
-        }
-    }
-
-    public void HealOverTimeRoutine(int value, int duration) {
-        healRoutineTimeLeft = duration;
-
-        if (healRoutine == null) {
-            healRoutine = StartCoroutine(HealOverTime(value));
-        }
-    }
-
-    IEnumerator HealOverTime(int value) {
-        while (healRoutineTimeLeft > 0) {
-            Debug.Log("Health: " + health);
-            Debug.Log("Time: " + healRoutineTimeLeft);
-            ModifyHealth(value);
-            if (health >= startingHealth) {
-                health = startingHealth;
-            } 
-            healRoutineTimeLeft -= 1f;
-            yield return new WaitForSeconds(1f);
-        }
-    }
-
-    public void DamageOverTimeRoutine(int value, int duration) {
-        damageRoutineTimeLeft = duration;
-
-        if (damageRoutine == null) {
-            damageRoutine = StartCoroutine(DamageOverTime(value));
-        }
-    }
-
-    IEnumerator DamageOverTime(int value) {
-        while (damageRoutineTimeLeft > 0) {
-            Debug.Log("Health: " + health);
-            Debug.Log("Time: " + damageRoutineTimeLeft);
-            ModifyHealth(value);
-            if (health <= 0) {
-                transform.position = new Vector2(0,0);
-                damageRoutine = null;
-                yield break;
-            }
-            damageRoutineTimeLeft -= 1f;
-            yield return new WaitForSeconds(1f);
         }
     }
 }
